Update libmseed to v3.5.1

+++ libmseed/packdata.c
@@ -312,7 +312,8 @@ msr_encode_steim1 (int32_t *input, uint64_t samplecount, int32_t *output, uint64
         /* Add new diffs and determine bit width needed to represent */
         for (idx = diffcount; idx < 4 && inputidx < (samplecount - 1); idx++, inputidx++)
         {
-          diffs[idx] = *(input + inputidx + 1) - *(input + inputidx);
+          /* Difference in unsigned to avoid signed overflow UB */
+          diffs[idx] = (int32_t) ((uint32_t) *(input + inputidx + 1) - (uint32_t) *(input + inputidx));
           BITWIDTH (diffs[idx], bitwidth[idx]);
           diffcount++;
         }
@@ -514,7 +515,8 @@ msr_encode_steim2 (int32_t *input, uint64_t samplecount, int32_t *output, uint64
         /* Add new diffs and determine bit width needed to represent */
         for (idx = diffcount; idx < 7 && inputidx < (samplecount - 1); idx++, inputidx++)
         {
-          diffs[idx] = *(input + inputidx + 1) - *(input + inputidx);
+          /* Difference in unsigned to avoid signed overflow UB */
+          diffs[idx] = (int32_t) ((uint32_t) *(input + inputidx + 1) - (uint32_t) *(input + inputidx));
           BITWIDTH (diffs[idx], bitwidth[idx]);
           diffcount++;
         }
